convert ICostRevaluationDAO to CostRevaluationRepository
